refactor generic HueInput
